Moved all OpenDroneID code to core.
