+++ Controls/NMEA_GPS_Connection.cs
