+++ Controls/NMEA_GPS_Connection.resx
