+++ MissionPlanner.sln
@@ -5,6 +5,7 @@ VisualStudioVersion = 17.3.32811.315
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MissionPlanner", "MissionPlanner.csproj", "{A2E22272-95FE-47B6-B050-9AE7E2055BF5}"
 	ProjectSection(ProjectDependencies) = postProject
+		{AF81C41B-06E6-4CA5-9FD4-A4E7042F1107} = {AF81C41B-06E6-4CA5-9FD4-A4E7042F1107}
 		{DF86AFC2-CA79-4137-9BB5-70AA529C3B79} = {DF86AFC2-CA79-4137-9BB5-70AA529C3B79}
 	EndProjectSection
 EndProject
@@ -34,28 +35,28 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SharpKml", "ExtLibs\SharpKm
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ZedGraph", "ExtLibs\ZedGraph\ZedGraph.csproj", "{2541686B-1673-43BF-AF89-3163945DB009}"
 	ProjectSection(ProjectDependencies) = postProject
-		{E4902330-74FE-45F6-86C3-0A5F73A1EEC0} = {E4902330-74FE-45F6-86C3-0A5F73A1EEC0}
 		{3566E9FD-591D-4E42-BA38-11DBE2F13E82} = {3566E9FD-591D-4E42-BA38-11DBE2F13E82}
+		{E4902330-74FE-45F6-86C3-0A5F73A1EEC0} = {E4902330-74FE-45F6-86C3-0A5F73A1EEC0}
 	EndProjectSection
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BSE.Windows.Forms", "ExtLibs\BSE.Windows.Forms\BSE.Windows.Forms.csproj", "{9CA367B8-0B98-49D1-84FB-735E612E3BA9}"
 	ProjectSection(ProjectDependencies) = postProject
-		{E4902330-74FE-45F6-86C3-0A5F73A1EEC0} = {E4902330-74FE-45F6-86C3-0A5F73A1EEC0}
 		{3566E9FD-591D-4E42-BA38-11DBE2F13E82} = {3566E9FD-591D-4E42-BA38-11DBE2F13E82}
+		{E4902330-74FE-45F6-86C3-0A5F73A1EEC0} = {E4902330-74FE-45F6-86C3-0A5F73A1EEC0}
 	EndProjectSection
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MetaDataExtractor", "ExtLibs\MetaDataExtractorCSharp240d\MetaDataExtractor.csproj", "{B46F6C6C-9184-41AF-8F8B-E0084752CA7C}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BaseClasses", "ExtLibs\BaseClasses\BaseClasses.csproj", "{2A8E8AF5-74E7-49DB-A42E-9360FA7A6CC4}"
 	ProjectSection(ProjectDependencies) = postProject
-		{E4902330-74FE-45F6-86C3-0A5F73A1EEC0} = {E4902330-74FE-45F6-86C3-0A5F73A1EEC0}
 		{3566E9FD-591D-4E42-BA38-11DBE2F13E82} = {3566E9FD-591D-4E42-BA38-11DBE2F13E82}
+		{E4902330-74FE-45F6-86C3-0A5F73A1EEC0} = {E4902330-74FE-45F6-86C3-0A5F73A1EEC0}
 	EndProjectSection
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MissionPlanner.Controls", "ExtLibs\Controls\MissionPlanner.Controls.csproj", "{C8B88795-6D01-494D-83AD-6944BD4C5023}"
 	ProjectSection(ProjectDependencies) = postProject
-		{E4902330-74FE-45F6-86C3-0A5F73A1EEC0} = {E4902330-74FE-45F6-86C3-0A5F73A1EEC0}
 		{3566E9FD-591D-4E42-BA38-11DBE2F13E82} = {3566E9FD-591D-4E42-BA38-11DBE2F13E82}
+		{E4902330-74FE-45F6-86C3-0A5F73A1EEC0} = {E4902330-74FE-45F6-86C3-0A5F73A1EEC0}
 	EndProjectSection
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MissionPlanner.Comms", "ExtLibs\Comms\MissionPlanner.Comms.csproj", "{825E7A10-390C-4A2B-B3A8-491D14966912}"
@@ -91,8 +92,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SharpAdbClient", "ExtLibs\S
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AltitudeAngelWings", "ExtLibs\AltitudeAngelWings\AltitudeAngelWings.csproj", "{098F542C-7C70-4B9C-B2E2-AC5DC5E0294B}"
 	ProjectSection(ProjectDependencies) = postProject
-		{E4902330-74FE-45F6-86C3-0A5F73A1EEC0} = {E4902330-74FE-45F6-86C3-0A5F73A1EEC0}
 		{3566E9FD-591D-4E42-BA38-11DBE2F13E82} = {3566E9FD-591D-4E42-BA38-11DBE2F13E82}
+		{E4902330-74FE-45F6-86C3-0A5F73A1EEC0} = {E4902330-74FE-45F6-86C3-0A5F73A1EEC0}
 	EndProjectSection
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "solo", "ExtLibs\solo\solo.csproj", "{CF402948-A3F4-4962-8A57-E40B46759D25}"
@@ -141,8 +142,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MissionPlanner.HIL", "ExtLi
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OSDConfigurator", "ExtLibs\OSDConfigurator\OSDConfigurator.csproj", "{1442B805-D6B1-4729-89F8-3A5A1C7547AD}"
 	ProjectSection(ProjectDependencies) = postProject
-		{E4902330-74FE-45F6-86C3-0A5F73A1EEC0} = {E4902330-74FE-45F6-86C3-0A5F73A1EEC0}
 		{3566E9FD-591D-4E42-BA38-11DBE2F13E82} = {3566E9FD-591D-4E42-BA38-11DBE2F13E82}
+		{E4902330-74FE-45F6-86C3-0A5F73A1EEC0} = {E4902330-74FE-45F6-86C3-0A5F73A1EEC0}
 	EndProjectSection
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DroneCAN", "ExtLibs\DroneCAN\DroneCAN.csproj", "{3A413E59-A20E-477C-997E-64B1398E0E80}"
@@ -220,8 +221,6 @@ EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NMEA2000", "ExtLibs\NMEA2000\NMEA2000.csproj", "{63F10356-A09B-46B0-9AFA-FAB8326A52DD}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AltitudeAngelWings.Plugin", "ExtLibs\AltitudeAngelWings.Plugin\AltitudeAngelWings.Plugin.csproj", "{DED183CA-5591-4CE5-A89D-987035BAC553}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OpenDroneID_Plugin", "Plugins\OpenDroneID2\OpenDroneID_Plugin.csproj", "{FBE658B3-684D-46E6-8668-50DEF0778538}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Bulb", "ExtLibs\LEDBulb\Bulb\Bulb.csproj", "{AF81C41B-06E6-4CA5-9FD4-A4E7042F1107}"
 EndProject
@@ -875,14 +874,6 @@ Global
 		{DED183CA-5591-4CE5-A89D-987035BAC553}.Release|Any CPU.Build.0 = Release|Any CPU
 		{DED183CA-5591-4CE5-A89D-987035BAC553}.Release|x86.ActiveCfg = Release|Any CPU
 		{DED183CA-5591-4CE5-A89D-987035BAC553}.Release|x86.Build.0 = Release|Any CPU
-		{FBE658B3-684D-46E6-8668-50DEF0778538}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{FBE658B3-684D-46E6-8668-50DEF0778538}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{FBE658B3-684D-46E6-8668-50DEF0778538}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{FBE658B3-684D-46E6-8668-50DEF0778538}.Debug|x86.Build.0 = Debug|Any CPU
-		{FBE658B3-684D-46E6-8668-50DEF0778538}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{FBE658B3-684D-46E6-8668-50DEF0778538}.Release|Any CPU.Build.0 = Release|Any CPU
-		{FBE658B3-684D-46E6-8668-50DEF0778538}.Release|x86.ActiveCfg = Release|Any CPU
-		{FBE658B3-684D-46E6-8668-50DEF0778538}.Release|x86.Build.0 = Release|Any CPU
 		{AF81C41B-06E6-4CA5-9FD4-A4E7042F1107}.Debug|Any CPU.ActiveCfg = Debug|x86
 		{AF81C41B-06E6-4CA5-9FD4-A4E7042F1107}.Debug|Any CPU.Build.0 = Debug|x86
 		{AF81C41B-06E6-4CA5-9FD4-A4E7042F1107}.Debug|x86.ActiveCfg = Debug|x86
@@ -975,7 +966,6 @@ Global
 		{6609C54A-71B9-4349-8E79-1DF3293BC7D0} = {0371886E-5A61-4626-B9E0-50A837ECE2DD}
 		{63F10356-A09B-46B0-9AFA-FAB8326A52DD} = {7E264F12-9EC3-4CDC-A187-6879C2B2BFCF}
 		{DED183CA-5591-4CE5-A89D-987035BAC553} = {9B3AC501-5FCB-4A2A-A628-BE0985DB2BD4}
-		{FBE658B3-684D-46E6-8668-50DEF0778538} = {9B3AC501-5FCB-4A2A-A628-BE0985DB2BD4}
 		{AF81C41B-06E6-4CA5-9FD4-A4E7042F1107} = {7E264F12-9EC3-4CDC-A187-6879C2B2BFCF}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
